fixed bug not allowing float inputs

+++ add.py
@@ -1,6 +1,6 @@
 def work(a, b):
-    a = int(a)
-    b = int(b)
+    a = float(a)
+    b = float(b)
     return [a + b, True]
 
 def run():
@@ -8,6 +8,6 @@ def run():
     nums = input("Write the numbers you want to add separated by spaces:\n")
     nums = nums.split(" ")
     for i in range(0, len(nums)):
-        current = int(nums[i])
+        current = float(nums[i])
         output += current
     print(output)
